feat: add generate types to scripts
