+++ packages/scripts/src/types.ts
@@ -1,6 +1,7 @@
 export enum Commands {
   'build' = 'build',
   'deploy' = 'deploy',
+  'types' = 'types',
   'run' = 'run',
 }
